+++ mlrun/runtimes/base.py
@@ -226,7 +226,7 @@ class BaseRuntime(ModelObj):
     def _get_k8s(self):
         # ideally we are not supposed to access k8s helper through client side
         # once separated client and server this if statement will be redundant
-        if self._is_api_server:
+        if is_running_as_api():
             import mlrun.api.utils.singletons.k8s
 
             return mlrun.api.utils.singletons.k8s.get_k8s_helper()
