+++ tests/api/runtimes/test_nuclio.py
@@ -28,23 +28,15 @@ from fastapi import HTTPException
 from fastapi.testclient import TestClient
 from sqlalchemy.orm import Session
 
+import mlrun.api.crud.runtimes.nuclio.function
 import mlrun.api.schemas
 import mlrun.errors
+import mlrun.runtimes.function
 import mlrun.runtimes.pod
 from mlrun import code_to_function, mlconf
 from mlrun.api.api.endpoints.functions import _build_function
 from mlrun.platforms.iguazio import split_path
 from mlrun.runtimes.constants import NuclioIngressAddTemplatedIngressModes
-from mlrun.runtimes.function import (
-    _compile_nuclio_archive_config,
-    compile_function_config,
-    deploy_nuclio_function,
-    enrich_function_with_ingress,
-    is_nuclio_version_in_range,
-    min_nuclio_versions,
-    resolve_function_ingresses,
-    validate_nuclio_version_compatibility,
-)
 from mlrun.utils import logger
 from tests.api.conftest import K8sSecretsMock
 from tests.api.runtimes.base import TestRuntimeBase
@@ -124,7 +116,9 @@ class TestNuclioRuntime(TestRuntimeBase):
         }
 
     def _execute_run(self, runtime, **kwargs):
-        deploy_nuclio_function(runtime, **kwargs)
+        mlrun.api.crud.runtimes.nuclio.function.deploy_nuclio_function(
+            runtime, **kwargs
+        )
 
     def _generate_runtime(
         self, kind=None, labels=None
@@ -371,7 +365,9 @@ class TestNuclioRuntime(TestRuntimeBase):
         function = self._generate_runtime(self.runtime_kind)
         key, val = "test.label.com/env", "test"
         function.set_label(key, val)
-        _, _, config = compile_function_config(function)
+        _, _, config = mlrun.api.crud.runtimes.nuclio.function._compile_function_config(
+            function
+        )
         assert config["metadata"]["labels"].get(key) == val
 
     def test_enrich_with_ingress_no_overriding(self, db: Session, client: TestClient):
@@ -384,12 +380,18 @@ class TestNuclioRuntime(TestRuntimeBase):
         # both ingress and node port
         ingress_host = "something.com"
         function.with_http(host=ingress_host, paths=["/"], port=30030)
-        function_name, project_name, config = compile_function_config(function)
+        (
+            function_name,
+            project_name,
+            config,
+        ) = mlrun.api.crud.runtimes.nuclio.function._compile_function_config(function)
         service_type = "NodePort"
-        enrich_function_with_ingress(
+        mlrun.api.crud.runtimes.nuclio.function._enrich_function_with_ingress(
             config, NuclioIngressAddTemplatedIngressModes.always, service_type
         )
-        ingresses = resolve_function_ingresses(config["spec"])
+        ingresses = mlrun.api.crud.runtimes.nuclio.function._resolve_function_ingresses(
+            config["spec"]
+        )
         assert len(ingresses) > 0, "Expected one ingress to be created"
         for ingress in ingresses:
             assert "hostTemplate" not in ingress, "No host template should be added"
@@ -400,12 +402,18 @@ class TestNuclioRuntime(TestRuntimeBase):
         Expect ingress template to be created as the configuration templated ingress mode is "always"
         """
         function = self._generate_runtime(self.runtime_kind)
-        function_name, project_name, config = compile_function_config(function)
+        (
+            function_name,
+            project_name,
+            config,
+        ) = mlrun.api.crud.runtimes.nuclio.function._compile_function_config(function)
         service_type = "NodePort"
-        enrich_function_with_ingress(
+        mlrun.api.crud.runtimes.nuclio.function._enrich_function_with_ingress(
             config, NuclioIngressAddTemplatedIngressModes.always, service_type
         )
-        ingresses = resolve_function_ingresses(config["spec"])
+        ingresses = mlrun.api.crud.runtimes.nuclio.function._resolve_function_ingresses(
+            config["spec"]
+        )
         assert ingresses[0]["hostTemplate"] != ""
 
     def test_enrich_with_ingress_on_cluster_ip(self, db: Session, client: TestClient):
@@ -414,14 +422,20 @@ class TestNuclioRuntime(TestRuntimeBase):
         function service type is ClusterIP
         """
         function = self._generate_runtime(self.runtime_kind)
-        function_name, project_name, config = compile_function_config(function)
+        (
+            function_name,
+            project_name,
+            config,
+        ) = mlrun.api.crud.runtimes.nuclio.function._compile_function_config(function)
         service_type = "ClusterIP"
-        enrich_function_with_ingress(
+        mlrun.api.crud.runtimes.nuclio.function._enrich_function_with_ingress(
             config,
             NuclioIngressAddTemplatedIngressModes.on_cluster_ip,
             service_type,
         )
-        ingresses = resolve_function_ingresses(config["spec"])
+        ingresses = mlrun.api.crud.runtimes.nuclio.function._resolve_function_ingresses(
+            config["spec"]
+        )
         assert ingresses[0]["hostTemplate"] != ""
 
     def test_enrich_with_ingress_never(self, db: Session, client: TestClient):
@@ -429,12 +443,18 @@ class TestNuclioRuntime(TestRuntimeBase):
         Expect no ingress to be created automatically as the configuration templated ingress mode is "never"
         """
         function = self._generate_runtime(self.runtime_kind)
-        function_name, project_name, config = compile_function_config(function)
+        (
+            function_name,
+            project_name,
+            config,
+        ) = mlrun.api.crud.runtimes.nuclio.function._compile_function_config(function)
         service_type = "DoesNotMatter"
-        enrich_function_with_ingress(
+        mlrun.api.crud.runtimes.nuclio.function._enrich_function_with_ingress(
             config, NuclioIngressAddTemplatedIngressModes.never, service_type
         )
-        ingresses = resolve_function_ingresses(config["spec"])
+        ingresses = mlrun.api.crud.runtimes.nuclio.function._resolve_function_ingresses(
+            config["spec"]
+        )
         assert ingresses == []
 
     def test_nuclio_config_spec_env(self, db: Session, client: TestClient):
@@ -457,7 +477,11 @@ class TestNuclioRuntime(TestRuntimeBase):
             {"name": name2, "value": value2},
         ]
 
-        function_name, project_name, config = compile_function_config(function)
+        (
+            function_name,
+            project_name,
+            config,
+        ) = mlrun.api.crud.runtimes.nuclio.function._compile_function_config(function)
         for expected_env_var in expected_env_vars:
             assert expected_env_var in config["spec"]["env"]
         assert isinstance(function.spec.env[0], kubernetes.client.V1EnvVar)
@@ -465,7 +489,11 @@ class TestNuclioRuntime(TestRuntimeBase):
 
         # simulating sending to API - serialization through dict
         function = function.from_dict(function.to_dict())
-        function_name, project_name, config = compile_function_config(function)
+        (
+            function_name,
+            project_name,
+            config,
+        ) = mlrun.api.crud.runtimes.nuclio.function._compile_function_config(function)
         for expected_env_var in expected_env_vars:
             assert expected_env_var in config["spec"]["env"]
 
@@ -996,43 +1024,75 @@ class TestNuclioRuntime(TestRuntimeBase):
     def test_is_nuclio_version_in_range(self):
         mlrun.runtimes.utils.cached_nuclio_version = "1.7.2"
 
-        assert not is_nuclio_version_in_range("1.6.11", "1.7.2")
-        assert not is_nuclio_version_in_range("1.7.0", "1.3.1")
-        assert not is_nuclio_version_in_range("1.7.3", "1.8.5")
-        assert not is_nuclio_version_in_range("1.7.2", "1.7.2")
-        assert is_nuclio_version_in_range("1.7.2", "1.7.3")
-        assert is_nuclio_version_in_range("1.7.0", "1.7.3")
-        assert is_nuclio_version_in_range("1.5.5", "1.7.3")
-        assert is_nuclio_version_in_range("1.5.5", "2.3.4")
+        assert not mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.6.11", "1.7.2"
+        )
+        assert not mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.7.0", "1.3.1"
+        )
+        assert not mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.7.3", "1.8.5"
+        )
+        assert not mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.7.2", "1.7.2"
+        )
+        assert mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.7.2", "1.7.3"
+        )
+        assert mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.7.0", "1.7.3"
+        )
+        assert mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.5.5", "1.7.3"
+        )
+        assert mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.5.5", "2.3.4"
+        )
 
         # best effort - assumes compatibility
         mlrun.runtimes.utils.cached_nuclio_version = ""
-        assert is_nuclio_version_in_range("1.5.5", "2.3.4")
-        assert is_nuclio_version_in_range("1.7.2", "1.7.2")
+        assert mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.5.5", "2.3.4"
+        )
+        assert mlrun.api.crud.runtimes.nuclio.function._is_nuclio_version_in_range(
+            "1.7.2", "1.7.2"
+        )
 
     def test_validate_nuclio_version_compatibility(self):
         # nuclio version we have
         mlconf.nuclio_version = "1.6.10"
 
-        # validate_nuclio_version_compatibility receives the min nuclio version required
-        assert not validate_nuclio_version_compatibility("1.6.11")
-        assert not validate_nuclio_version_compatibility("1.5.9", "1.6.11")
-        assert not validate_nuclio_version_compatibility("1.6.11", "1.5.9")
-        assert not validate_nuclio_version_compatibility("2.0.0")
-        assert validate_nuclio_version_compatibility("1.6.9")
-        assert validate_nuclio_version_compatibility("1.5.9")
+        # mlrun.runtimes.function.validate_nuclio_version_compatibility receives the min nuclio version required
+        assert not mlrun.runtimes.function.validate_nuclio_version_compatibility(
+            "1.6.11"
+        )
+        assert not mlrun.runtimes.function.validate_nuclio_version_compatibility(
+            "1.5.9", "1.6.11"
+        )
+        assert not mlrun.runtimes.function.validate_nuclio_version_compatibility(
+            "1.6.11", "1.5.9"
+        )
+        assert not mlrun.runtimes.function.validate_nuclio_version_compatibility(
+            "2.0.0"
+        )
+        assert mlrun.runtimes.function.validate_nuclio_version_compatibility("1.6.9")
+        assert mlrun.runtimes.function.validate_nuclio_version_compatibility("1.5.9")
 
         mlconf.nuclio_version = "2.0.0"
-        assert validate_nuclio_version_compatibility("1.6.11")
-        assert validate_nuclio_version_compatibility("1.5.9", "1.6.11")
+        assert mlrun.runtimes.function.validate_nuclio_version_compatibility("1.6.11")
+        assert mlrun.runtimes.function.validate_nuclio_version_compatibility(
+            "1.5.9", "1.6.11"
+        )
 
         # best effort - assumes compatibility
         mlconf.nuclio_version = ""
-        assert validate_nuclio_version_compatibility("1.6.11")
-        assert validate_nuclio_version_compatibility("1.5.9", "1.6.11")
+        assert mlrun.runtimes.function.validate_nuclio_version_compatibility("1.6.11")
+        assert mlrun.runtimes.function.validate_nuclio_version_compatibility(
+            "1.5.9", "1.6.11"
+        )
 
         with pytest.raises(ValueError):
-            validate_nuclio_version_compatibility("")
+            mlrun.runtimes.function.validate_nuclio_version_compatibility("")
 
     def test_min_nuclio_versions_decorator_failure(self):
         mlconf.nuclio_version = "1.6.10"
@@ -1043,7 +1103,7 @@ class TestNuclioRuntime(TestRuntimeBase):
             ["1.5.9", "1.6.11"],
         ]:
 
-            @min_nuclio_versions(*case)
+            @mlrun.runtimes.function.min_nuclio_versions(*case)
             def fail():
                 pytest.fail("Should not enter this function")
 
@@ -1060,7 +1120,7 @@ class TestNuclioRuntime(TestRuntimeBase):
                 ["1.0.0", "0.9.81", "1.4.1"],
             ]:
 
-                @min_nuclio_versions(*case)
+                @mlrun.runtimes.function.min_nuclio_versions(*case)
                 def success():
                     pass
 
@@ -1296,7 +1356,11 @@ class TestNuclioRuntime(TestRuntimeBase):
         if build_secret_name is not None:
             fn.spec.build.secret = build_secret_name
 
-        _, _, deployed_config = compile_function_config(fn)
+        (
+            _,
+            _,
+            deployed_config,
+        ) = mlrun.api.crud.runtimes.nuclio.function._compile_function_config(fn)
         assert deployed_config["spec"].get("imagePullSecrets") == expected_secret_name
 
     def test_nuclio_with_preemption_mode(self):
@@ -1476,11 +1540,19 @@ class TestNuclioRuntime(TestRuntimeBase):
 
         if expected_ingress_host_template is None:
             # never
-            ingresses = resolve_function_ingresses(deploy_spec)
+            ingresses = (
+                mlrun.api.crud.runtimes.nuclio.function._resolve_function_ingresses(
+                    deploy_spec
+                )
+            )
             assert ingresses == []
 
         else:
-            ingresses = resolve_function_ingresses(deploy_spec)
+            ingresses = (
+                mlrun.api.crud.runtimes.nuclio.function._resolve_function_ingresses(
+                    deploy_spec
+                )
+            )
             assert ingresses[0]["hostTemplate"] == expected_ingress_host_template
 
 
@@ -1495,6 +1567,8 @@ class TestNuclioMLRunRuntime(TestNuclioRuntime):
 def get_archive_spec(function, secrets):
     spec = nuclio.ConfigSpec()
     config = {}
-    _compile_nuclio_archive_config(spec, function, secrets)
+    mlrun.api.crud.runtimes.nuclio.function._compile_nuclio_archive_config(
+        spec, function, secrets
+    )
     spec.merge(config)
     return config
